初步支持emcc (#38)

+++ test/test_filesystem.lua
@@ -5,6 +5,7 @@ local shell = require 'shell'
 local supported = require 'supported'
 
 local isWindows = platform.os == 'windows'
+local isEmscripten = platform.os == 'emscripten'
 local isMinGW   = isWindows and platform.CRT == 'libstdc++'
 local isMacOS   = platform.os ~= "macos"
 
@@ -515,7 +516,7 @@ function test_fs:test_rename()
     fs.remove_all(fs.path('temp2'))
     fs.create_directories(fs.path('temp1'))
     fs.create_directories(fs.path('temp2'))
-    if isWindows then
+    if isWindows or isEmscripten then
         rename_failed('temp1', 'temp2')
     else
         rename_ok('temp1', 'temp2')
@@ -529,8 +530,10 @@ function test_fs:test_rename()
     rename_failed('temp1', 'temp2')
 end
 
-function test_fs:test_current_path()
-    lt.assertEquals(fs.current_path(), fs.path(shell:pwd()))
+if supported "popen" then
+    function test_fs:test_current_path()
+        lt.assertEquals(fs.current_path(), fs.path(shell:pwd()))
+    end
 end
 
 function test_fs:test_copy_file()
@@ -772,32 +775,34 @@ end
 --    assertPathEquals(fs.dll_path(), fs.absolute(getdll()))
 --end
 
-function test_fs:test_filelock_1()
-    local lock = fs.path("temp.lock")
-    local f1 = lt.assertIsUserdata(fs.filelock(lock))
-    lt.assertEquals(fs.filelock(lock), nil)
-    f1:close()
-    local f2 = lt.assertIsUserdata(fs.filelock(lock))
-    f2:close()
-    fs.remove(fs.path("temp.lock"))
-end
+if platform.os ~= 'emscripten' then
+    function test_fs:test_filelock_1()
+        local lock = fs.path("temp.lock")
+        local f1 = lt.assertIsUserdata(fs.filelock(lock))
+        lt.assertEquals(fs.filelock(lock), nil)
+        f1:close()
+        local f2 = lt.assertIsUserdata(fs.filelock(lock))
+        f2:close()
+        fs.remove(fs.path("temp.lock"))
+    end
 
-function test_fs:test_filelock_2()
-    local process = shell:runlua([[
+    function test_fs:test_filelock_2()
+        local process = shell:runlua([[
         local fs = require 'bee.filesystem'
         fs.filelock(fs.path("temp.lock"))
         io.stdout:write 'ok'
         io.stdout:flush()
         io.read 'a'
     ]], { stdin = true, stdout = true, stderr = true })
-    lt.assertEquals(process.stdout:read(2), 'ok')
-    lt.assertEquals(fs.filelock(fs.path("temp.lock")), nil)
-    process.stdin:close()
-    lt.assertEquals(process.stderr:read 'a', '')
-    lt.assertEquals(process:wait(), 0)
-    local f = lt.assertIsUserdata(fs.filelock(fs.path("temp.lock")))
-    f:close()
-    fs.remove(fs.path("temp.lock"))
+        lt.assertEquals(process.stdout:read(2), 'ok')
+        lt.assertEquals(fs.filelock(fs.path("temp.lock")), nil)
+        process.stdin:close()
+        lt.assertEquals(process.stderr:read 'a', '')
+        lt.assertEquals(process:wait(), 0)
+        local f = lt.assertIsUserdata(fs.filelock(fs.path("temp.lock")))
+        f:close()
+        fs.remove(fs.path("temp.lock"))
+    end
 end
 
 function test_fs:test_tostring()
